enable startup and liveness probes for active-web-elements-server
The server has started when all the kernels have started, and is live so
long as each kernel pool has running kernels.

Then, because the .applicationserver/info and
.applicationserver/kernel/stats endpoints don't need kernels, they
should be accessible even if the server isn't otherwise ready.  This
makes it possible to use ServiceConnect to deploy nodefiles, including
the initial mathpass, to the server as part of server configuration.